Split overview modules for combineReducers: (#64)
combineReducers expects that a reducer exists for each part of the state
it operates on, otherwise it console.errors that a key is in the state
that it doesn't know about. Also without a module listed in
combineReducers the initial state of the store is set up with any extra
state as undefined. This causes prop validation errors when initialising
the component as isOverviewOn is undefined at that point.

The later could be fixed by setting a default in the aggregations
component but does not solve the first annoying issue. This is an
in-between solution that allows isOverviewOn to continue handling the
toggle event in the root module, while allowing for the is-overview-on
module to hold the action creator and the reducer that sets up the
initial state.